# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Agromart B2B agricultural marketplace backend server.",
        "title": "Agromart API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/update/image/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns an S3 presigned URL to upload a profile image for the given admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Image Upload"
                ],
                "summary": "Generates presigned URL for admin profile image upload and updates the image path in database",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PresignedURLResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid URL Param",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/business/update/image/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns an S3 presigned URL to upload a profile image for the given business",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Image Upload"
                ],
                "summary": "Generate presigned URL for business profile image upload and updates the image path in database",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PresignedURLResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid URL Param",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/category/sub/update/image/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns an S3 presigned URL to upload an image for the given sub-category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Image Upload"
                ],
                "summary": "Generate presigned URL for sub-category image upload and updates the image path in database",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PresignedURLResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid URL Param",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/category/update/image/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns an S3 presigned URL to upload an image for the given category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Image Upload"
                ],
                "summary": "Generate presigned URL for category image upload and updates the image path in database",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PresignedURLResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid URL Param",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/product/delete/image": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns success message after successfully deleting image from both S3 and database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Image Upload"
                ],
                "summary": "Delete's the Product Image from S3 and image path from database",
                "parameters": [
                    {
                        "description": "Update Product Image Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.productImageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "product image deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid payload or missing fields",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/product/update/image": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns an S3 presigned URL to upload an image for the given product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Image Upload"
                ],
                "summary": "Generate presigned URL for product image upload and updates the image path in database",
                "parameters": [
                    {
                        "description": "Update Product Image Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.productImageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PresignedURLResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload or missing fields",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/update/image/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns an S3 presigned URL to upload a profile image for the given user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Image Upload"
                ],
                "summary": "Generate presigned URL for user profile image upload and updates the image path in database",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PresignedURLResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid URL Param",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
...
        "handlers.PresignedURLResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "presigned url generated successfully"
                },
                "url": {
                    "type": "string",
                    "example": "https://s3.amazonaws.com/bucket/profile/admins/uuid.png?X-Amz-Signature=..."
                }
            }
        },
...
        "handlers.productImageRequest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}